Extracts `expect` helper method to shared scope to remove duplication.
